Rename foundation to spud
