--- source/library/foundation/support/CMakeLists.txt
@@ -1,2 +1,0 @@
-include(up_target_natvis)
-up_target_natvis(foundation "${CMAKE_CURRENT_SOURCE_DIR}/up_foundation.natvis")
