+++ source/library/spud/public/potato/spud/fixed_string_writer.h
