+++ source/library/spud/support/CMakeLists.txt
@@ -1,0 +1,2 @@
+include(up_target_natvis)
+up_target_natvis(spud "${CMAKE_CURRENT_SOURCE_DIR}/up_spud.natvis")
