+++ source/library/spud/tests/test_hash.cpp
@@ -1,7 +1,7 @@
-#include "potato/foundation/hash.h"
-#include "potato/foundation/string_view.h"
-#include "potato/foundation/zstring_view.h"
-#include "potato/foundation/span.h"
+#include "potato/spud/hash.h"
+#include "potato/spud/string_view.h"
+#include "potato/spud/zstring_view.h"
+#include "potato/spud/span.h"
 #include <doctest/doctest.h>
 
 DOCTEST_TEST_SUITE("[potato][foundation] up::hash") {
