+++ source/library/spud/tests/test_string.cpp
@@ -1,4 +1,4 @@
-#include "potato/foundation/string.h"
+#include "potato/spud/string.h"
 #include <doctest/doctest.h>
 #include <ostream>
