+++ source/library/spud/tests/test_zstring_view.cpp
@@ -1,4 +1,4 @@
-#include "potato/foundation/zstring_view.h"
+#include "potato/spud/zstring_view.h"
 #include <doctest/doctest.h>
 #include <ostream>
 #include <cstring>
